refactor(runtime): narrow distance helpers to f32

+++ crimson-zig/src/runtime/helpers.zig
@@ -36,13 +36,13 @@ pub fn withinNativeFindRadius(
     return margin < 0.0;
 }
 
-pub fn distanceSq(a: state_mod.Vec2, b: state_mod.Vec2) f64 {
+pub fn distanceSq(a: state_mod.Vec2, b: state_mod.Vec2) f32 {
     const dx = a.x - b.x;
     const dy = a.y - b.y;
     return dx * dx + dy * dy;
 }
 
-pub fn distanceSqRoundedF32(a: state_mod.Vec2, b: state_mod.Vec2) f64 {
+pub fn distanceSqRoundedF32(a: state_mod.Vec2, b: state_mod.Vec2) f32 {
     return narrowF32(distanceSq(a, b));
 }
